requirements.txt
numpy
tensorflow
scikit-learn
matplotlib
pillow